requirements.txt
# I listed the Python packages I needed for this project here.
# Installing this file sets up training, generation, and evaluation dependencies.
torch
transformers
datasets
accelerate
evaluate
rouge-score
bert-score
sacrebleu
numpy
pyyaml
tqdm